+++ src/lib/pages/contract-details/index.tsx
@@ -35,7 +35,7 @@ import { useDerivedWasmVerifyInfo } from "lib/services/verification/wasm";
 import type { BechAddr32 } from "lib/types";
 import { jsonPrettify, truncate } from "lib/utils";
 
-import { CommandSection } from "./components/CommandSection";
+import { CommandSection } from "./components/command-section";
 import { ContractDesc } from "./components/contract-description";
 import { ContractStates } from "./components/contract-states";
 import { ContractTop } from "./components/ContractTop";
@@ -76,7 +76,7 @@ const ContractDetailsBody = observer(
     const { isTotalBondedLoading, totalBonded } =
       useAccountDelegationInfos(contractAddress);
 
-    const { data: relatedWasmVerifyInfo } = useDerivedWasmVerifyInfo(
+    const { data: derivedWasmVerifyInfo } = useDerivedWasmVerifyInfo(
       contractData?.contract.codeId,
       contractData?.contract.codeHash
     );
@@ -121,7 +121,7 @@ const ContractDetailsBody = observer(
           publicInfo={publicInfo}
           contract={contract}
           contractLocalInfo={contractLocalInfo}
-          wasmVerifyInfo={relatedWasmVerifyInfo}
+          wasmVerifyInfo={derivedWasmVerifyInfo}
         />
         <Tabs
           index={Object.values(TabIndex).indexOf(tab)}
@@ -171,15 +171,16 @@ const ContractDetailsBody = observer(
                     />
                   )}
                   <ContractVerificationSection
+                    contractAddress={contract.address}
                     codeId={contract.codeId}
                     codeHash={contract.codeHash}
-                    wasmVerifyInfo={relatedWasmVerifyInfo}
-                    contractAddress={contract.address}
+                    wasmVerifyInfo={derivedWasmVerifyInfo}
                   />
                   <CommandSection
                     contractAddress={contractAddress}
-                    codeHash={contract.codeHash}
                     codeId={contract.codeId}
+                    codeHash={contract.codeHash}
+                    wasmVerifyInfo={derivedWasmVerifyInfo}
                   />
                   <Flex
                     borderBottom={{ base: "0px", md: "1px solid" }}
@@ -228,7 +229,7 @@ const ContractDetailsBody = observer(
                         contract={contract}
                         contractRest={contractRest}
                         codeLocalInfo={codeLocalInfo}
-                        wasmVerifyInfo={relatedWasmVerifyInfo}
+                        wasmVerifyInfo={derivedWasmVerifyInfo}
                       />
                       <Button
                         size="sm"