+++ src/lib/pages/publish-module/components/UploadModuleCard.tsx
@@ -9,7 +9,7 @@ import { ComponentLoader } from "lib/components/ComponentLoader";
 import { DropZone } from "lib/components/dropzone";
 import { CustomIcon } from "lib/components/icon";
 import { Tooltip } from "lib/components/Tooltip";
-import { UploadCard } from "lib/components/upload/UploadCard";
+import { UploadCard } from "lib/components/upload";
 import { useDecodeModule } from "lib/services/move/module";
 import type { DecodeModuleQueryResponse } from "lib/services/types";
 import type { Option, UpgradePolicy } from "lib/types";