+++ src/lib/services/verification/wasm/index.ts
@@ -1,4 +1,4 @@
-import { useMutation, useQuery } from "@tanstack/react-query";
+import { useMutation, useQuery, useQueryClient } from "@tanstack/react-query";
 
 // TODO: refactor app-provider/app-fns
 import { useCelatoneApp } from "lib/app-provider/contexts";
@@ -11,43 +11,54 @@ import {
   submitWasmVerify,
 } from "./api";
 
-export const useSubmitWasmVerify = () =>
-  useMutation({
+export const useSubmitWasmVerify = () => {
+  const queryClient = useQueryClient();
+  const { currentChainId } = useCelatoneApp();
+
+  return useMutation({
     mutationFn: submitWasmVerify,
+    onSuccess(data, variables) {
+      queryClient.invalidateQueries({
+        queryKey: [
+          CELATONE_QUERY_KEYS.WASM_VERIFICATION_INFOS,
+          currentChainId,
+          variables.codeId,
+        ],
+      });
+    },
   });
+};
 
 export const useWasmVerifyInfos = (codeIds: number[], enabled = true) => {
   const { currentChainId } = useCelatoneApp();
-  return useQuery(
-    [
+  return useQuery({
+    queryKey: [
       CELATONE_QUERY_KEYS.WASM_VERIFICATION_INFOS,
       currentChainId,
       ...codeIds.sort(),
     ],
-    () => getWasmVerifyInfos(currentChainId, codeIds),
-    {
-      enabled,
-      refetchOnWindowFocus: false,
-      retry: false,
-    }
-  );
+    queryFn: () => getWasmVerifyInfos(currentChainId, codeIds),
+    enabled,
+    refetchOnWindowFocus: false,
+    retry: false,
+    staleTime: Infinity,
+  });
 };
 
 const useWasmRelatedVerifyInfos = (hashes: string[], enabled = true) => {
   const { currentChainId } = useCelatoneApp();
-  return useQuery(
-    [
-      CELATONE_QUERY_KEYS.WASM_VERIFICATION_INFOS,
+  return useQuery({
+    queryKey: [
+      CELATONE_QUERY_KEYS.WASM_RELATED_VERIFICATION_INFOS,
       currentChainId,
       ...hashes.sort(),
     ],
-    () => getWasmRelatedVerifyInfos(currentChainId, hashes),
-    {
-      enabled,
-      refetchOnWindowFocus: false,
-      retry: false,
-    }
-  );
+    queryFn: () => getWasmRelatedVerifyInfos(currentChainId, hashes),
+    enabled,
+    refetchOnWindowFocus: false,
+    retry: false,
+    staleTime: Infinity,
+  });
 };
 
 export const useDerivedWasmVerifyInfo = (
@@ -65,7 +76,8 @@ export const useDerivedWasmVerifyInfo = (
     return { data: undefined, isLoading: true };
 
   const wasmVerifyInfo = wasmVerifyInfos?.[Number(codeId)];
-  const wasmRelatedVerifyInfo = wasmRelatedVerifyInfos?.[hash ?? ""];
+  const wasmRelatedVerifyInfo =
+    wasmRelatedVerifyInfos?.[hash?.toUpperCase() ?? ""];
   return {
     data: {
       verificationInfo: wasmVerifyInfo?.verificationInfo ?? null,